some initial ui setup and resource uploads
