+++ app/src/main/java/com/cs446group18/delaywise/ui/flightinfo/FlightInfoViewModel.kt
@@ -1,0 +1,13 @@
+package com.cs446group18.delaywise.ui.flightinfo
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class FlightInfoViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is gallery Fragment"
+    }
+    val text: LiveData<String> = _text
+}
