+++ app/src/main/java/com/cs446group18/delaywise/ui/home/HomeFragment.kt
@@ -28,10 +28,7 @@ class HomeFragment : Fragment() {
         _binding = FragmentHomeBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
-            textView.text = it
-        }
+        homeViewModel.text.observe(viewLifecycleOwner) {}
         return root
     }
