+++ app/src/main/java/com/cs446group18/delaywise/ui/home/HomeViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class HomeViewModel : ViewModel() {
-
     private val _text = MutableLiveData<String>().apply {
         value = "This is home Fragment"
     }
