+++ app/src/main/java/com/cs446group18/delaywise/ui/savedflights/SavedFlightsViewModel.kt
@@ -1,0 +1,13 @@
+package com.cs446group18.delaywise.ui.savedflights
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SavedFlightsViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is slideshow Fragment"
+    }
+    val text: LiveData<String> = _text
+}
