+++ app/src/main/res/drawable/__plane_icon.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="37dp"
+    android:height="33dp"
+    android:viewportWidth="37"
+    android:viewportHeight="33">
+  <path
+      android:pathData="M36.929,16.719C36.913,14.859 34.382,13.203 31.132,13.203H24.507C23.632,13.203 23.319,13.016 22.772,12.453L12.007,1C11.648,0.625 11.241,0.438 10.788,0.438H8.476C8.069,0.438 7.819,0.813 8.022,1.234L13.616,13.172L5.491,14.063L2.585,9.063C2.351,8.672 2.007,8.5 1.507,8.5H0.772C0.319,8.5 0.007,8.813 0.007,9.25V24.188C0.007,24.625 0.319,24.938 0.772,24.938H1.507C2.007,24.938 2.351,24.75 2.585,24.375L5.491,19.375L13.616,20.266L8.022,32.188C7.819,32.625 8.069,33 8.476,33H10.788C11.241,33 11.648,32.813 12.007,32.438L22.772,20.984C23.319,20.406 23.632,20.234 24.507,20.234H31.132C34.382,20.234 36.913,18.578 36.929,16.719Z"
+      android:fillColor="#1B33B4"/>
+</vector>
