+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -9,12 +9,12 @@
             android:icon="@drawable/ic_menu_camera"
             android:title="@string/menu_home" />
         <item
-            android:id="@+id/nav_gallery"
+            android:id="@+id/nav_flightinfo"
             android:icon="@drawable/ic_menu_gallery"
-            android:title="@string/menu_gallery" />
+            android:title="@string/menu_flightinfo" />
         <item
-            android:id="@+id/nav_slideshow"
+            android:id="@+id/nav_savedflights"
             android:icon="@drawable/ic_menu_slideshow"
-            android:title="@string/menu_slideshow" />
+            android:title="@string/menu_savedflights" />
     </group>
 </menu>
