+++ app/src/main/res/values/strings.xml
@@ -2,12 +2,12 @@
     <string name="app_name">DelayWise</string>
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="nav_header_title">Android Studio</string>
-    <string name="nav_header_subtitle">android.studio@android.com</string>
+    <string name="nav_header_title">Bob Smith</string>
+    <string name="nav_header_subtitle">bob.smith@xyz.com</string>
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
 
     <string name="menu_home">Home</string>
-    <string name="menu_gallery">Gallery</string>
-    <string name="menu_slideshow">Slideshow</string>
+    <string name="menu_flightinfo">Flight Info</string>
+    <string name="menu_savedflights">Saved Flights</string>
 </resources>
